requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
openai == 0.28.0
